Use autoloaded functions & move ui to after.
Changed functions to be autoloaded. Cool feature.
Tweaked commands to use the autoloaded functions correctly.

Moved ui to after, not actually sure if I need to do that or not.
My thinking was that I want to be sure that if a colorscheme is getting
loaded as a plugin, that it wasn't going to be used until *after* it has
been loaded.